requirements.txt
amqp==2.1.1
appdirs==1.4.3
Babel==2.3.4
beautifulsoup4==4.6.0
billiard==3.5.0.2
boto3==1.4.2
botocore==1.4.81
celery==4.0.0
click==6.6
coverage==4.3.4
defusedxml==0.5.0rc1
Django==2.2.3
django-admin-rangefilter==0.3.7
django-appconf==1.0.2
django-autocomplete-light==3.2.1
django-cors-headers==1.3.1
django-db-mailer==2.4.0
django-debug-toolbar==2.0
django-guardian==1.4.9
django-imagekit==3.3
django-jet==1.0.8
django-push-notifications==1.6.0
django-storages==1.5.1
django-uuidfield==0.5.0
djangorestframework==3.9.4
docutils==0.12
drf-tracking==1.5.0
factory-boy==2.11.1
Faker==0.9.2
fcm-django==0.2.19
Flask==0.11.1
flower==0.9.1
freezegun==0.3.8
gevent==1.1.2
greenlet==0.4.10
h2==2.5.1
hpack==2.3.0
hyper==0.7.0
hyperframe==3.2.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kombu==4.0.0
livereload==2.5.1
locustio==0.8a2
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==0.23
mkdocs==0.17.3
mockredispy==2.9.3
msgpack-python==0.4.8
oauthlib==2.0.0
olefile==0.45.1
packaging==16.8
pilkit==1.1.13
Pillow==4.2.1
psycopg2==2.6.2
pyfcm==1.4.5
PyJWT==1.4.2
pyparsing==2.2.0
python-dateutil==2.6.0
python-social-auth==0.3.6
python3-openid==3.0.10
pytz==2016.7
PyYAML==3.12
pyzmq==15.2.0
redis==2.10.5
requests==2.21.0
requests-oauthlib==0.7.0
requests-toolbelt==0.8.0
s3transfer==0.1.9
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.2.2
text-unidecode==1.2
tornado==4.2
vine==1.1.3
Werkzeug==0.11.11
